Separating expiration from deletion. Adding BackupExpirationMonitor and BackupSweeper

+++ mbs/retention.py
@@ -1,12 +1,11 @@
 __author__ = 'abdul'
 
-import traceback
 import logging
 import persistence
 import operator
+import traceback
 
 from mbs import get_mbs
-from errors import RetentionPolicyError
 
 from base import MBSObject
 from date_utils import date_now, date_minus_seconds
@@ -14,7 +13,7 @@ from date_utils import date_now, date_minus_seconds
 
 from schedule_runner import ScheduleRunner
 from schedule import Schedule
-from task import STATE_SUCCEEDED, STATE_CANCELED, STATE_FAILED
+from task import STATE_SUCCEEDED
 
 from task import EVENT_TYPE_ERROR
 from target import CloudBlockStorageSnapshotReference
@@ -39,38 +38,6 @@ class RetentionPolicy(MBSObject):
     ###########################################################################
     def __init__(self):
         MBSObject.__init__(self)
-
-    ###########################################################################
-
-    def apply_policy(self, plan):
-        """
-            Applies the retention policy by deleting target references for
-            expired succeeded backups (e.g. deleting backup files for expired
-            backups)
-        """
-        policy_name = self.__class__.__name__
-
-        for backup in self.get_backups_due(plan):
-            try:
-                expire_backup(backup)
-
-            except Exception, e:
-                logger.error("%s: Error while archiving backup %s. "
-                             "Trace: %s" %
-                             (policy_name, backup.id, traceback.format_exc()))
-
-                msg = ("Error while applying retention policy on backup %s. " %
-                       backup.id)
-                raise RetentionPolicyError(msg, cause=e,
-                                           details=traceback.format_exc())
-
-    ###########################################################################
-    def get_backups_due(self, plan):
-        q = _backups_to_check_query(plan_id=plan.id)
-
-        backups = get_mbs().backup_collection.find(query=q)
-
-        return self.filter_backups_due_for_expiration(backups)
 
     ###########################################################################
     def filter_backups_due_for_expiration(self, backups):
@@ -171,35 +138,45 @@ class RetainMaxTimePolicy(RetentionPolicy):
         }
 
 ###############################################################################
-# BackupSweeper
+# BackupExpirationMonitor
 ###############################################################################
 
-DEFAULT_SWEEP_SCHEDULE = Schedule(frequency_in_seconds=20 * 60 * 60)
+DEFAULT_MONITOR_SCHEDULE = Schedule(frequency_in_seconds=(2 * 60 * 60))
 
 
-class BackupSweeper(ScheduleRunner):
-
+class BackupExpirationMonitor(ScheduleRunner):
+    """
+        A Thread that periodically expire backups that are due for expiration
+    """
     ###########################################################################
     def __init__(self, schedule=None):
-        schedule = schedule or DEFAULT_SWEEP_SCHEDULE
+        schedule = schedule or DEFAULT_MONITOR_SCHEDULE
         ScheduleRunner.__init__(self, schedule=schedule)
-        logger.info("Initializing BackupSweeper")
+        logger.info("Initializing BackupExpirationMonitor")
 
     ###########################################################################
     def tick(self):
-        self._expire_backups_due()
+        try:
+            self._expire_backups_due()
+        except Exception, ex:
+            logger.exception("BackupExpirationMonitor Error")
+            subject = "BackupExpirationMonitor Error"
+            message = ("BackupExpirationMonitor Error!.\n\nStack Trace:\n%s" %
+                       traceback.format_exc())
+            get_mbs().send_error_notification(subject, message, ex)
 
     ###########################################################################
     def _expire_backups_due(self):
 
-        logger.info("BackupSweeper: Starting a sweep cycle...")
+        logger.info("BackupExpirationMonitor: Starting an expiration check "
+                    "cycle...")
         total_processed = 0
         total_expired = 0
         total_dont_expire = 0
 
-        logger.info("BackupSweeper: Finding all recurring backups due for "
-                    "expiration")
-        q = self.get_backups_to_check_query()
+        logger.info("BackupExpirationMonitor: Finding all recurring backups"
+                    " due for expiration")
+        q = _check_to_expire_query()
 
         q["plan._id"] = {
             "$exists": True
@@ -207,7 +184,7 @@ class BackupSweeper(ScheduleRunner):
 
         s = [("plan._id", -1)]
 
-        logger.info("BackupSweeper: Executing query :\n%s" %
+        logger.info("BackupExpirationMonitor: Executing query :\n%s" %
                     document_pretty_string(q))
 
         backups_iter = get_mbs().backup_collection.find_iter(query=q, sort=s)
@@ -237,16 +214,16 @@ class BackupSweeper(ScheduleRunner):
                 plan_backups = []
 
         # process onetime backups
-        logger.info("BackupSweeper: Finding all onetime backups due for "
-                    "expiration")
+        logger.info("BackupExpirationMonitor: Finding all onetime backups "
+                    "due for expiration")
 
-        q = self.get_backups_to_check_query()
+        q = _check_to_expire_query()
 
         q["plan._id"] = {
             "$exists": False
         }
 
-        logger.info("BackupSweeper: Executing query :\n%s" %
+        logger.info("BackupExpirationMonitor: Executing query :\n%s" %
                     document_pretty_string(q))
         onetime_backups_iter = get_mbs().backup_collection.find_iter(query=q)
 
@@ -259,9 +236,9 @@ class BackupSweeper(ScheduleRunner):
                 mark_backup_never_expire(current_backup)
                 total_dont_expire += 1
 
-
-        logger.info("BackupSweeper: Finished sweeping cycle. Total Expired=%s,"
-                    " Total Dont Expire=%s, Total Processed=%s" %
+        logger.info("BackupExpirationMonitor: Finished expiration check cycle. "
+                    "Total Expired=%s, Total Don't Expire=%s, "
+                    "Total Processed=%s" %
                     (total_expired, total_dont_expire, total_processed))
 
     ###########################################################################
@@ -298,106 +275,205 @@ class BackupSweeper(ScheduleRunner):
 
         return len(dues) if dues else 0
 
-    ###########################################################################
-    def get_backups_to_check_query(self):
-        return _backups_to_check_query()
 
+###############################################################################
+# BackupSweeper
+###############################################################################
+
+DEFAULT_SWEEP_SCHEDULE = Schedule(frequency_in_seconds=12 * 60 * 60)
+
+
+class BackupSweeper(ScheduleRunner):
+    """
+        A Thread that periodically deletes backups targets that
+        are due for deletion
+    """
+    ###########################################################################
+    def __init__(self, schedule=None):
+        schedule = schedule or DEFAULT_SWEEP_SCHEDULE
+        ScheduleRunner.__init__(self, schedule=schedule)
+        logger.info("Initializing BackupSweeper")
+
+    ###########################################################################
+    def tick(self):
+        try:
+            self._delete_backups_targets_due()
+        except Exception, ex:
+            logger.exception("BackupSweeper Error")
+            subject = "BackupSweeper Error"
+            message = ("BackupSweeper Error!.\n\nStack Trace:\n%s" %
+                       traceback.format_exc())
+            get_mbs().send_error_notification(subject, message, ex)
+
+    ###########################################################################
+    def _delete_backups_targets_due(self):
+
+        logger.info("BackupSweeper: Starting a sweep cycle...")
+        total_processed = 0
+        total_deleted = 0
+        total_errored = 0
+
+        logger.info("BackupSweeper: Finding all backups"
+                    " due for deletion")
+        q = _check_to_delete_query()
+
+        q["plan._id"] = {
+            "$exists": True
+        }
+
+        s = [("plan._id", -1)]
+
+        logger.info("BackupSweeper: Executing query :\n%s" %
+                    document_pretty_string(q))
+
+        backups_iter = get_mbs().backup_collection.find_iter(query=q, sort=s)
+
+        # process all plan backups
+        for backup in backups_iter:
+            total_processed += 1
+            try:
+                delete_backup_targets(backup)
+                total_deleted += 1
+            except Exception, ex:
+                logger.exception("BackupSweeper: Error while attempting to "
+                                 "delete backup targets for backup '%s'" %
+                                 backup.id)
+                subject = "BackupSweeper Error"
+                message = ("BackupSweeper Error!.\n\nStack Trace:\n%s" %
+                            traceback.format_exc())
+                get_mbs().send_error_notification(subject, message, ex)
+                total_errored += 1
+
+        logger.info("BackupSweeper: Finished sweep cycle. "
+                    "Total Deleted=%s, Total Errored=%s, "
+                    "Total Processed=%s" %
+                    (total_deleted, total_errored, total_processed))
 
 ###############################################################################
 # QUERY HELPER
 ###############################################################################
-def _backups_to_check_query(plan_id=None):
+def _check_to_expire_query():
     q = {
-        "targetReference": {
-            "$exists": True
-        },
+        "state": STATE_SUCCEEDED,
         "expiredDate": {"$exists": False},
         "dontExpire": {"$ne": True}
     }
 
-    if plan_id:
-        q["plan._id"] = plan_id
+    return q
+
+###############################################################################
+def _check_to_delete_query():
+    """
+        We only delete backups that got expired at least two days ago.
+        This is just to make sure that if the expiration monitor screws up we
+         would still have time to see what happened
+    """
+    two_days_ago = date_minus_seconds(date_now(), 2 * 24 * 60 * 60)
+    q = {
+        "expiredDate": {
+            "$lt": two_days_ago
+        },
+        "deletedDate": {
+            "$exists": False
+        }
+    }
 
     return q
 
 ###############################################################################
-# EXPIRE BACKUP HELPERS
+# EXPIRE/DELETE BACKUP HELPERS
 ###############################################################################
-def expire_backup(backup, expired_date=None, force=False):
-    try:
-        expired_date = expired_date or date_now()
-        return robustified_expire_backup(backup, expired_date, force=force)
-    except Exception, e:
-        msg = "Error while attempting to expire backup '%s': " % e
-        logger.exception(msg)
-        persistence.update_backup(backup, event_name="EXPIRE_ERROR",
-                                  message=msg, event_type=EVENT_TYPE_ERROR)
-        # if the backup expiration has errored out for 3 times then mark as
-        # unexpirable
-        if backup.event_logged_count("EXPIRE_ERROR") >= 3:
-            logger.info("Giving up on expire backup '%s'. Failed at least"
-                        " three times. Need to mark as unexpirable" %
-                        backup.id)
-            mark_backup_never_expire(backup)
-            return False
-        else:
-            raise
+def expire_backup(backup):
 
-
-###############################################################################
-@robustify(max_attempts=3, retry_interval=5,
-           do_on_exception=raise_if_not_retriable,
-           do_on_failure=raise_exception)
-def robustified_expire_backup(backup, expired_date, force=False):
-    """
-        expires the backup
-    """
 
     # do some validation
     if not backup.target_reference:
         raise BackupDeleteError("Cannot expire backup '%s'. "
                                 "Backup never uploaded" % backup.id)
 
-    # validate if the backup has been expired already
-    if backup.expired_date:
-        logger.warning("expire_backup(): Backup '%s' is already expired."
-                       " Ignoring..." % backup.id)
-        return
-
     # validate backups is expirable now if its part of a retained plan
-    if not force and backup.plan and backup.plan.retention_policy:
+    if backup.plan and backup.plan.retention_policy:
         validate_backup_should_expire_now(backup)
 
-    logger.info("Expiring backup '%s', expired date: '%s'." %
-                (backup.id, expired_date))
+    try:
+        backup.expired_date = date_now()
+        persistence.update_backup(backup, properties="expiredDate",
+                                  event_name="EXPIRING", message="Expiring")
+
+    except Exception, e:
+        msg = "Error while attempting to expire backup '%s': " % e
+        logger.exception(msg)
+
+###############################################################################
+def delete_backup_targets(backup):
+    try:
+        robustified_delete_backup(backup)
+    except Exception, e:
+        msg = "Error while attempting to expire backup '%s': " % e
+        logger.exception(msg)
+        persistence.update_backup(backup, event_name="DELETE_ERROR",
+                                  message=msg, event_type=EVENT_TYPE_ERROR)
+        # if the backup expiration has errored out for 3 times then mark as
+        # unexpirable
+        #if backup.event_logged_count("DELETE_ERROR") >= 3:
+         #   logger.info("Giving up on delete backup '%s'. Failed at least"
+          #              " three times. Marking backup as deleted" %
+           #             backup.id)
+
+            #return False
+        #else:
+        raise
+
+
+###############################################################################
+@robustify(max_attempts=3, retry_interval=5,
+           do_on_exception=raise_if_not_retriable,
+           do_on_failure=raise_exception)
+def robustified_delete_backup(backup):
+    """
+        deletes the backup targets
+    """
+
+    # do some validation
+    if not backup.target_reference:
+        raise BackupDeleteError("Cannot delete backup '%s'. "
+                                "Backup never uploaded" % backup.id)
+
+    logger.info("Deleting target references for backup '%s'." % backup.id)
 
     target_ref = backup.target_reference
 
-    #TODO re-enable actual target deletion for all cases (not limited to force)
-    exists = True
-    # if the target reference is a cloud storage one then make the cloud
-    # storage object take care of it
-    if force:
-        logger.info("Deleting actual backup target since temp flag force=true "
-                    "for '%s'" % backup.id)
-        exists = do_delete_target_ref(backup, target_ref)
+    logger.info("Deleting primary target reference for backup '%s'." %
+                backup.id)
+    do_delete_target_ref(backup, backup.target, target_ref)
 
-        # expire log file
-        if backup.log_target_reference:
-            pass
-            exists = do_delete_target_ref(backup, backup.log_target_reference)
-    else:
-        logger.info("Marking backup '%s' with expiredDate WITHOUT deleting "
-                    "actual target (temporary)" % backup.id)
-    # set expired date
-    backup.expired_date = expired_date
-    backup.expired_for_reals = force
-    persistence.update_backup(backup, properties=["expiredDate",
-                                                  "expiredForReals"],
-                              event_name="EXPIRING", message="Expiring")
+    # delete log file
+    if backup.log_target_reference:
+        logger.info("Deleting log target reference for backup '%s'." %
+                    backup.id)
+        do_delete_target_ref(backup, backup.target,
+                             backup.log_target_reference)
 
-    logger.info("Backup %s expired successfully!" % backup.id)
-    return exists
+    if backup.secondary_target_references:
+        logger.info("Deleting secondary target references for backup '%s'." %
+                    backup.id)
+        sec_targets = backup.secondary_targets
+        sec_target_refs = backup.secondary_target_references
+        for (sec_target, sec_tgt_ref) in zip(sec_targets, sec_target_refs):
+            logger.info("Deleting secondary target reference %s for backup "
+                        "'%s'." % (sec_tgt_ref, backup.id))
+            do_delete_target_ref(backup, sec_target, sec_tgt_ref)
+
+    # set deleted date
+    backup.deleted_date = date_now()
+    update_props = ["deletedDate", "targetReference",
+                    "secondaryTargetReferences"]
+    persistence.update_backup(backup, properties=update_props,
+                              event_name="DELETING",
+                              message="Deleting target references")
+
+    logger.info("Backup %s target references deleted successfully!" %
+                backup.id)
 
 ###############################################################################
 def validate_backup_should_expire_now(backup):
@@ -430,10 +506,19 @@ def mark_backup_never_expire(backup):
                               message="Marking as dontExpire")
 
 ###############################################################################
-def do_delete_target_ref(backup, target_ref):
+def do_delete_target_ref(backup, target, target_ref):
+
+    if target_ref.preserve:
+        logger.info("Skipping deletion for target ref %s (backup '%s') because"
+                    " it is preserved" % (target_ref, backup.id))
+        return
+
+    target_ref.deleted_date = date_now()
+    # if the target reference is a cloud storage one then make the cloud
+    # storage object take care of it
     if isinstance(target_ref, CloudBlockStorageSnapshotReference):
         logger.info("Deleting backup '%s' snapshot " % backup.id)
         return target_ref.cloud_block_storage.delete_snapshot(target_ref)
     else:
         logger.info("Deleting backup '%s file" % backup.id)
-        return backup.target.delete_file(target_ref)
+        return target.delete_file(target_ref)
